backend/requirements.txt
Django
djangorestframework
textblob
django-cors-headers
celery
redis
waitress
